refactoring component hierarchy
