+++ src/components/forms/base-fields/option-fields/delete.vue
