+++ src/components/forms/base-fields/option-fields/options/delete.vue
